day10/part2: finish part2 using dynamic programming
A simple approach similar to how the part1 was redone, with the help of
a hashmap, and building the solution from partial ones, using dynamic
programming.

To elaborate, the idea is to find the complete number of combinations
possible. We realize that if we look at the number of combinations for a
given jolt value J, those would be composed of the combinations from
adapters with jolt value J - 1, J - 2, and J - 3, since these are all
the reachable adapters. This entails that a solution starting from the
first (0) adapter which is only reachable by itself (1), can be built
from there until the maximum which is the max seen in the input + 3 (the
built in adapter jolt value).

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>