Add day15 part1
I was tricked by the newline inside the 'input' file!
The buffer to which I was reading the stdin had a new line at the end
and I only noticed it after running and passing all the unit tests and
still not having the right answer.

What happens is that split(',').filter_map(|x| x.parse()), actually
doesn't parse correctly the last number with a newline!! All the rest of
the numbers didn't have a space around the ',' so I didn't notice it.

Pay to make assertions on the assumptions early on!!

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>